+++ frb_example/pure_dart/dart/lib/main.dart
@@ -193,15 +193,26 @@ void main(List<String> args) async {
     expect(names, ['Steve', 'Bob', 'Alex']);
   });
 
-  test('dart call handleComplexStruct', () async {
-    final arrLen = 5;
-    final complexStructResp = await api.handleComplexStruct(s: _createMyTreeNode(arrLen: arrLen));
+  testComplexStruct(MyTreeNode complexStructResp, {required int arrLen}) {
     expect(complexStructResp.valueI32, 100);
     expect(complexStructResp.valueVecU8, List.filled(arrLen, 100));
     expect(complexStructResp.children[0].valueVecU8, List.filled(arrLen, 110));
     expect(complexStructResp.children[0].children[0].valueVecU8, List.filled(arrLen, 111));
     expect(complexStructResp.children[1].valueVecU8, List.filled(arrLen, 120));
+  }
+
+  test('dart call handleComplexStruct', () async {
+    final arrLen = 5;
+    final complexStructResp = await api.handleComplexStruct(s: _createMyTreeNode(arrLen: arrLen));
+    testComplexStruct(complexStructResp, arrLen: arrLen);
   });
+
+  test('dart call handleNestedStruct', () async {
+    final r = await api.handleNestedStruct(s: _createMyNestedStruct());
+    testComplexStruct(r.treeNode, arrLen: 5);
+    expect(r.weekday, Weekdays.Friday);
+  });
+
   test('dart call handleComplexStructSync', () {
     final arrLen = 5;
     final complexStructResp = api.handleComplexStructSync(s: _createMyTreeNode(arrLen: arrLen));
@@ -677,6 +688,26 @@ void main(List<String> args) async {
       final resp = await api.duration(d: duration);
       expect(resp.inHours, duration.inHours);
     });
+
+    test('Combined Chrono types', () async {
+      final test = await api.testChrono();
+      expect(castInt(test.dt!.millisecondsSinceEpoch), castInt(1631297333000));
+      expect(castInt(test.dt2!.millisecondsSinceEpoch), castInt(1631297333000));
+      expect(test.du, Duration(hours: 4));
+    });
+
+    test('combined chrono types precise', () async {
+      final datetime_1 = DateTime.utc(2002, 02, 23, 12, 13, 55);
+      final datetime_2 = DateTime.utc(1800, 01, 23, 12, 56, 25);
+      final duration = Duration(hours: 4);
+
+      final result = await api.testPreciseChrono();
+
+      expect(result.dt!.millisecondsSinceEpoch, datetime_1.millisecondsSinceEpoch);
+      expect(result.dt2!.millisecondsSinceEpoch, datetime_2.millisecondsSinceEpoch);
+      expect(result.du!.inHours, duration.inHours);
+    });
+
     test('nested chrono types', () async {
       const duration = Duration(hours: 4);
       final naive = DateTime.utc(2022, 09, 10, 20, 48, 53, 123, 456);
@@ -1178,6 +1209,10 @@ MyTreeNode _createMyTreeNode({required int arrLen}) {
       ),
     ],
   );
+}
+
+MyNestedStruct _createMyNestedStruct() {
+  return MyNestedStruct(treeNode: _createMyTreeNode(arrLen: 5), weekday: Weekdays.Friday);
 }
 
 class MatchBigInt extends CustomMatcher {