+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.65.0
 
 * Support empty structs #1071 (thanks @alexthe2)
+* Fix Optional<DateTime> #1079 (thanks @alexthe2)
 
 ## 1.64.0